requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==22.0.0  # https://github.com/benoitc/gunicorn
psycopg[c]==3.1.18  # https://github.com/psycopg/psycopg

# Django
# ------------------------------------------------------------------------------
# django-anymail[sendgrid]==10.3  # https://github.com/anymail/django-anymail
# JWT
djangorestframework-simplejwt

# passowrd reset
django-rest-passwordreset

# email
django-sendgrid-v5